requirements.txt
albumentations==0.4.6
anyio==3.4.0
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
branca==0.4.2
certifi==2021.10.8
cffi==1.14.3
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
folium==0.12.1
fonttools==4.28.5
h11==0.12.0
httptools==0.2.0
idna==3.3
imageio==2.13.5
imgaug==0.4.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mistune==0.8.4
multipledispatch==0.6.0
natsort==8.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.6.3
notebook==6.1.4
numpy==1.21.5
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
orjson==3.6.5
packaging==20.4
pandas==1.3.5
pandas-flavor==0.2.0
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
plotly-express==0.4.1
predicthq==1.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.7.1
pyjanitor==0.21.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.26.0
schematics==2.0.0.dev2
scikit-image==0.19.1
scipy==1.7.3
Send2Trash==1.5.0
Shapely==1.8.0
six==1.15.0
sniffio==1.2.0
starlette==0.16.0
statsmodels==0.13.1
tenacity==8.0.1
terminado==0.9.1
testpath==0.4.4
tifffile==2021.11.2
timezonefinder==5.2.0
tonyg-rfc3339==0.1
torch==1.10.1
torchvision==0.11.2
tornado==6.0.4
traitlets==5.0.5
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.5.1
xarray==0.20.2
